Rearrange pyframe stuff to match CPython structure
