+++ pyo3-ffi/src/pytypedefs.rs
@@ -1,0 +1,5 @@
+// TODO: Created this file as part of fixing pyframe.rs and cpython/pyframe.rs
+// TODO: Finish defining or moving declarations now in Include/pytypedefs.h
+
+#[cfg(Py_LIMITED_API)]
+opaque_struct!(pub PyFrameObject);
